+++ R/browse.R
@@ -57,7 +57,8 @@ NULL
 #' @export
 #' @rdname browse-this
 browse_package <- function(package = NULL) {
-  stopifnot(is.null(package) || is_string(package))
+  maybe_name(package)
+
   if (is.null(package)) {
     check_is_project()
   }
@@ -156,7 +157,7 @@ browse_cran <- function(package = NULL) {
 # 2. BugReports/URL fields of DESCRIPTION (active project or arbitrary
 #    installed package)
 github_url <- function(package = NULL) {
-  stopifnot(is.null(package) || is_string(package))
+  maybe_name(package)
 
   if (is.null(package)) {
     check_is_project()