+++ tests/testthat/test-version.R
@@ -23,11 +23,16 @@ test_that("use_version() increments version in DESCRIPTION, edits NEWS", {
     value = "1.1.1.9000",
     overwrite = TRUE
   )
+  mock_cran_version("1.1.1")
   use_news_md()
 
   use_version("major")
   expect_identical(proj_version(), "2.0.0")
-  expect_match(read_utf8(proj_path("NEWS.md"), n = 1), "2.0.0")
+
+  expect_snapshot(
+    writeLines(read_utf8(proj_path("NEWS.md"))),
+    transform = scrub_testpkg
+  )
 })
 
 test_that("use_dev_version() appends .9000 to Version, exactly once", {
@@ -42,6 +47,7 @@ test_that("use_dev_version() appends .9000 to Version, exactly once", {
 test_that("use_version() updates (development version) directly", {
   create_local_package()
   use_description_field(name = "Version", value = "0.0.1", overwrite = TRUE)
+  mock_cran_version("0.0.1")
   use_news_md()
 
   # bump to dev to set (development version)
@@ -50,14 +56,9 @@ test_that("use_version() updates (development version) directly", {
   # directly overwrite development header
   use_version("patch")
 
-  expect_match(
-    read_utf8(proj_path("NEWS.md"), n = 1),
-    "0[.]0[.]2"
-  )
-
-  expect_match(
-    read_utf8(proj_path("NEWS.md"), n = 3)[3],
-    "0[.]0[.]1"
+  expect_snapshot(
+    writeLines(read_utf8(proj_path("NEWS.md"))),
+    transform = scrub_testpkg
   )
 })
 
@@ -78,5 +79,16 @@ test_that("use_version() updates version.c", {
   use_dev_version()
 
   lines <- read_utf8(ver_path)
-  expect_true(grepl("1.0.0.9000", lines, fixed = TRUE)[[2]])
+  expect_snapshot(writeLines(lines), transform = scrub_testpkg)
+})
+
+test_that("is_dev_version() detects dev version directly and with DESCRIPTION", {
+  expect_true(is_dev_version("0.0.1.9000"))
+  expect_false(is_dev_version("0.0.1"))
+
+  create_local_package()
+  use_description_field(name = "Version", value = "1.0.0", overwrite = TRUE)
+  expect_false(is_dev_version())
+  use_dev_version()
+  expect_true(is_dev_version())
 })